backend/comp333-homework2-django/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
psycopg2-binary==2.8.6
python-dateutil==1.5
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
